Implement ships using the SRTS Expanded mod
Both the Tel'tal and Wraith Dart are fully working. I think there's still some references to the old texture and Helicopter stuff in the C# source that I don't have yet, so it does spew some console errors on start, but everything works nonetheless.

This does make the mod depend on SRTS Expanded.